refactor: move pkg/messagepool/messagesigner.go to pkg/messagesigner/messagesigner.go
